Some reorganization of the code. Functionally only change is when the node is declared as dead, we clear out the initialization statistics as well forcing reestablishment of the reverse link .

+++ MicroFrameworkPK_v4_3/DeviceCode/pal/COM/Wireless/MAC/OMAC/OMACTimeSync.cpp
@@ -305,24 +305,27 @@ DeviceStatus OMACTimeSync::Receive(RadioAddress_t msg_src, TimeSyncMsg* rcv_msg,
 			}
 		}
 
-		//For each time table entry, if the neighbor is not active in neighbor table, kick it from both the neighbor table and the time table
-		for(UINT8 i =0; i < MAX_NBR; ++i){
-			if(INVALID_NBR_ID != g_OMAC.m_omac_scheduler.m_TimeSyncHandler.m_globalTime.regressgt2.samples[i].nbrID){
-				if( g_NeighborTable.FindIndexEvenDead( g_OMAC.m_omac_scheduler.m_TimeSyncHandler.m_globalTime.regressgt2.samples[i].nbrID, &index) == DS_Success){
-					if(g_NeighborTable.Neighbor[index].neighborStatus != Alive){ //If it exists in the neighbortable but marked as not alive, delete from the time entires from globalTime Table
-						g_NeighborTable.ClearNeighborwIndex(index);
-						g_NeighborTable.NumberValidNeighbor ++ ;
-						//g_NeighborTable.Neighbor[index].Clear();
-						g_OMAC.m_omac_scheduler.m_TimeSyncHandler.m_globalTime.regressgt2.Clean(g_OMAC.m_omac_scheduler.m_TimeSyncHandler.m_globalTime.regressgt2.samples[i].nbrID);
-						if(!m_globalTime.regressgt2.IsTableFull()) {
-							is_space_available_in_table = true;
-							m_globalTime.regressgt2.InsertNbrID(msg_src);
-							break;
+		if(!is_space_available_in_table){ //If still space not available look for stale entries in the neighbor table
+			//For each time table entry, if the neighbor is not active in neighbor table, kick it from both the neighbor table and the time table
+			for(UINT8 i =0; i < MAX_NBR; ++i){
+				if(INVALID_NBR_ID != g_OMAC.m_omac_scheduler.m_TimeSyncHandler.m_globalTime.regressgt2.samples[i].nbrID){
+					if( g_NeighborTable.FindIndexEvenDead( g_OMAC.m_omac_scheduler.m_TimeSyncHandler.m_globalTime.regressgt2.samples[i].nbrID, &index) == DS_Success){
+						if(g_NeighborTable.Neighbor[index].neighborStatus != Alive){ //If it exists in the neighbortable but marked as not alive, delete from the time entires from globalTime Table
+							g_NeighborTable.ClearNeighborwIndex(index);
+							g_NeighborTable.NumberValidNeighbor ++ ;
+							//g_NeighborTable.Neighbor[index].Clear();
+							g_OMAC.m_omac_scheduler.m_TimeSyncHandler.m_globalTime.regressgt2.Clean(g_OMAC.m_omac_scheduler.m_TimeSyncHandler.m_globalTime.regressgt2.samples[i].nbrID);
+							if(!m_globalTime.regressgt2.IsTableFull()) {
+								is_space_available_in_table = true;
+								m_globalTime.regressgt2.InsertNbrID(msg_src);
+								break;
+							}
 						}
 					}
 				}
 			}
 		}
+
 	}
 
 	if(is_space_available_in_table){
